requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==4.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
drf-nested-routers==0.93.5
idna==3.6
mysqlclient==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==1.7.1
pymongo==3.12.3
PyMySQL==1.1.0
pytz==2023.4
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.2.4
tzdata==2023.4
urllib3==2.2.1